Try the extra package caches before the main one
When an admin adds an extra package cache to install/update packages, it
is because they want to use the packages from this extra cache in
priority.

Adding them after the main one means that the following could happen:

1. the package is fully downloaded in the extra cache,
2. the package had been partially downloaded in the main cache,
3. installing the package will try finishing the download from the main
   cache instead of just using the full one from the extra cache.

We obviously do not want that, as the goal of the extra package cache
mechanism is to avoid downloading stuff as much as possible.

Fixes #821