static: Use a custom storage
Back in the days, we were using Django's default static files storage.
This was very inefficient because we weren't telling clients to cache
the static files, so the poor Koombooks were crawling under the load
created by all those static file requests.

So we decided to configure Nginx to cache the static files forever, but
that meant we had to find a way to change the URL whenever the file
changed.

That's why we started using Django's ManifestStaticFilesStorage, which:

* copies the `filename.extension` files to `filename.filemd5.extension`
  when running `collectstatic`;
* replaces the URLs to the ones with the file MD5;

And that worked well, except when it didn't: for no apparent reason the
manifest file would sometimes be empty, leading to 500 errors.

We tried to figure out why this manifest file was sometimes (quite often
on Koombooks in fact) empty, but just couldn't find the root cause.

But taking a step back, all we want is to cache the static files
forever, and change the URL whenever the file changes.

One approximation to that is to use the Ideascube version in the URL, so
that the URL changes whenever Ideascube is updated. It's a bit worse
(new requests get fired the first time a static file is loaded after an
Ideascube update, even if the file hasn't changed), but it's still
pretty good, especially since most servers never get any update.

And we can implement that trivially with our own custom storage class.

It also means that `collectstatic` is much faster, since it doesn't need
to compute the MD5 of all the static files. The poor Koombooks will
appreciate, especially as we run `collectstatic` when starting the app.

Overall this seems like a pretty good compromise.

Fixes #829