catalog: Fix a test
The point of this test was to ensure that explicitly upgrading a package
which isn't installed will just install it.

Instead, what it was doing was asking to update all installed packages.

(that succeeded though, because of another issue: instead of upgrading
all installed packages we were upgrading all **available** packages...
see a future commit)