Store StockItem.physical in the database
Two of our views list the physical stock items: the stock list view and
the inventory update view.

Until now, we would have to actually get all the stock items, iterate on
them, and for each one:

* if it is in fact a Book, then find if is has any "physical" book
specimen:
  * if it does, then this stock item is physical
  * if it does not, then this stock item is not physical

* if it is not, then this stock item is physical

Testing with a database containing 514 stock items and 1479 specimens, we
end up running 1238 SQL queries just to get this list!

It is so terribly slow that on a Koombook, Nginx ends up timing out because
uWSGI takes too long to respond.

This all started when we made the Book model inherit from the StockItem one,
back in !542. We absolutely didn't plan for this explosion of database queries,
and it's now biting us back.

This commit is the simplest possibility to fix this, making the
StockItem.physical attribute a field which persists in the database.

Suddenly listing the physical stock items can be done in a single query.

However, it introduces the possibility for discrepancies in our data:
when adding a physical book specimen to a book, we need to make sure we
set the corresponding Book.physical attribute to True.

We also need to ensure that existing physical stock items remain
physical after this change, which the migration tries to do.

Fixes #841