Remove the wikipedia and gutenberg services
Those don't exist any more. A while ago we moved to having a single
Kiwix server, but we forgot to remove those services.

This is causing much confusion, with box admins telling us that
"Wikipedia and Gutenberg don't work", when they in fact do, but those
services appear in error.