Remove the kalite service
The main use case for this was to let admins stop the service when they
don't use it, because itused to be extremely resource-hungry.

The Kalite folks made a lot of progress on this though, to the point
that these days we just install it when needed, and then nobody feels
the need to stop it.