Remove the services page
Now that the only two things left are the Nginx and uWSGI services,
we've decided to remove the page entirely.

Letting people stop any of these two services is akin to letting them
shoot themselves in the foot: after doing that the only thing they can
do is a hard poweroff of the machine.

And restarting those services doesn't really bring them anything more
than a clean reboot would do.

As a result, let's just remove this whole page, and all the underlying
code.

Fixes #557
Fixes #839