Change the way we are installing ZippedZim.
Until now, we were extracting all files in 'data' directory of the
zipped-zim and use the internal/individual `library.xml` to generate the
global `library.xml`.

But, as we want to move to simple zim, the `library.xml` will not exist.
So, instead of using `library.xml` to pass meta-information to the
`Kiwix` handler, create our own metadata file.

The `Kiwix` handle still need to look for `library.xml` files as we still
need to use zim files installed the "old" way.

When a zim will be updated/reinstalled, it will be re-installed the new
way.

As there is no more `date` entry in the metadata files, there will be no
`date` entry in the `library.xml`. To test that the package has been
actually updated, we need to check for the checksum. So we also need
to have a real different data. That's why the wikipedia.tum-2015-09
is changed.