Be sure that we correctly load cache yml file and convert it to json.
If we had a value than could be parsed as a date by yaml
(`2014-10-30` instead of `'2014-10-30'`), yaml parsed it as a
datetime object.

First we don't want a datetime object at all as we are waiting for str.
Second, it crashes when we try to dump the same data in json as
json cannot handle "complex" python object.