Fix broken json migration because of wrong datetime parsed by yaml.
As the write to the json cache file may have been broken because of
value being parsed as datetime by yaml, we may ends with malformed json
files.

Hopefully, in this case, the yml file is still there. So use it to get
the data intead of the json file.