Stop using the date in the library.xml to check that update is done.
With simple zim installation, we will not have a associated `library.xml`
and so, the book in the global `library.xml` will not have a `date`
attribute.

So, checking the date in the library.xml seems to no be a good test.
We now test the sha256sum of the installed zim to be sure that the new
file has been actually installed.

We also need to have real different data if we want different checksum.
The wikipedia.tum-2015-09.zip is change this way.
A `\0` byte has been added to the end of the zim file in the zip package.
It will create invalid zim file, but we will have a different checksum.