blog: Create articles with the user language by default
Currently the language selected by default when writing a new article is
the default UI language, as configured in the settings file.

But in most cases, people write in the language they read.

That is, it is highly likely that if a staff browses Ideascube in
Spanish, then they will also write articles in Spanish.

And in cases this assumption is wrong, then this is not a worse default
than the current one anyway.

This is based on the discussion we had on @Lucas-C's merge request #875
who initially started working on this.

Fixes #760
Closes #875