Make the shutdown button look more dangerous
Shutting down the server is a destructive operation, as it will not come
back up on its own. Someone will need to physically go to the server,
and press the power button.

We have been told that users find the form confusing and hard to
understand, some even clicking on the wrong button.

Hopefully this helps quickly distinguishing the two different actions.

Fixes #838