Add some more details to the stock export
Feedback from the field told us that staff would like to have the serial
and barcode in the export. I'm adding the count, because I have a strong
feeling the request would come next anyway.

This requires changing the model on which the export operates, to avoid
running too many queries.