Allow branding of the css.
In pibox-installer we need to be able to customize the css
(ideascube/pibox-installer#122)

By loading the the `style.css` file in the `branding` directory/url, we
would be able to customize the css as we are customizing logo and favicon.

We also need a empty `style.css` to avoid 404 request.