search: Only display the right filters
Currently the sidebar displays all filters, even those that would lead
nowhere.

This is because the sidebar does queries using the search system, while
the list of elements (e.g blog articles, mediacenter documents, ...)
uses direct queries on the related model.

For example, if we had two blog articles:

* a draft in English;
* a published article in French;

Currently a non-staff user would see only the published article. However
the sidebar would display filter options for both languages. If the user
clicks on English, they get an empty list since no published article is
in English.

The same applies to tags, and document kinds: we should only display
filters in the sidebar for elements that can be viewed.

This commit achieves that by telling the search system to respect the
`public` index, which itself corresponds to whether the element model is
visible by non-staff. (e.g for the blog, `public` is true for published
articles only)