search: Only display the right filters
We were displaying all filters, even those that would lead to nowhere.

For example, if we had two blog articles:

* a draft in English;
* a published article in French;

The the sidebar would display both languages, even if a patron would get
an empty page when clicking on the "English" filter.

The same goes for tags in the blog, as well as document kinds, languages
and tags.

In the case of the mediacenter though, even staff shouldn't see hidden
documents.