Try to force the download of media by the browser on "download" button.
When a user click on the "download" button, he's expecting the file to
be downloaded. However, the browser try to display it.

By using the the `download` attribute to a link, we can ask the browser
to always download the file.
This is not always supported (https://caniuse.com/#feat=download) but
hopefully it will be enough.

Fix #868