blog: Don't show medias in the automatic summary
When computing automatic summaries (in case the author didn't write one
themselves), we truncate the text, preserving the overall HTML
structure.

However, when an article has a media close to the start, the media gets
integrated into the summary, which makes the list of articles look
pretty awful.

This commit strips the medias (img, video, audio, ...) when computing
the summary.

Fixes #867