Fix sidebar form going over the page footer.
On wide screen, as the sidebar is a float, if it is pretty long it can go
over the footer.

By adding a last element at the end of the content with `clear:both`
attribute, we force the container element to contain the whole sidebar.
And the footer is always after the container element, so after the sidebar.