conf: Improve ALLOWED_HOSTS for local access
We had 'localhost' allowed for all boxes by default, but 127.0.0.1 only
for development.

The two being equivalent, there's no reason not to define them at the
same place.

Local access to the app is not something we should worry about, so let's
just allow it by default in all cases.